+++ module/hooks/render-chat-message.mjs
@@ -3,5 +3,6 @@ import { BRPChat } from '../apps/chat.mjs'
 export function listen() {
   Hooks.on('renderChatMessageHTML', (app, html, data) => {
     BRPChat.renderMessageHook(app, html, data)
+    ui.chat.scrollBottom();
   })
 }